ai_freshness_app/requirements.txt
Flask>=3.0,<4.0
Pillow
scikit-learn
numpy
joblib
